Add Payment field to BaseExecutableContent with specifics for payment type
