+++ aleph_message/models/__init__.py
@@ -1,59 +1,19 @@
 import datetime
 import json
 from copy import copy
-from enum import Enum
 from hashlib import sha256
 from json import JSONDecodeError
 from pathlib import Path
-from typing import (
-    Any,
-    Dict,
-    List,
-    Literal,
-    Optional,
-    Type,
-    Union, TypeVar, cast,
-)
+from typing import Any, Dict, List, Literal, Optional, Type, TypeVar, Union, cast
 
 from pydantic import BaseModel, Extra, Field, validator
 from typing_extensions import TypeAlias
 
 from .abstract import BaseContent
+from .base import Chain, HashType, MessageType
 from .execution.instance import InstanceContent
 from .execution.program import ProgramContent
 from .item_hash import ItemHash, ItemType
-
-
-class Chain(str, Enum):
-    """Supported chains"""
-
-    AVAX = "AVAX"
-    BSC = "BSC"
-    CSDK = "CSDK"
-    DOT = "DOT"
-    ETH = "ETH"
-    NEO = "NEO"
-    NULS = "NULS"
-    NULS2 = "NULS2"
-    SOL = "SOL"
-    TEZOS = "TEZOS"
-
-
-class HashType(str, Enum):
-    """Supported hash functions"""
-
-    sha256 = "sha256"
-
-
-class MessageType(str, Enum):
-    """Message types supported by Aleph"""
-
-    post = "POST"
-    aggregate = "AGGREGATE"
-    store = "STORE"
-    program = "PROGRAM"
-    instance = "INSTANCE"
-    forget = "FORGET"
 
 
 class MongodbId(BaseModel):
@@ -205,7 +165,9 @@ class BaseMessage(BaseModel):
     size: Optional[int] = Field(
         default=None, description="Size of the content"
     )  # Almost always present
-    time: datetime.datetime = Field(description="Unix timestamp or datetime when the message was published")
+    time: datetime.datetime = Field(
+        description="Unix timestamp or datetime when the message was published"
+    )
     item_type: ItemType = Field(description="Storage method used for the content")
     item_content: Optional[str] = Field(
         default=None,
@@ -344,7 +306,7 @@ AlephMessage: TypeAlias = Union[
 ]
 
 
-T = TypeVar('T', bound=AlephMessage)
+T = TypeVar("T", bound=AlephMessage)
 
 AlephMessageType: TypeAlias = Type[T]
