+++ aleph_message/models/execution/base.py
@@ -1,6 +1,10 @@
 from __future__ import annotations
 
 from enum import Enum
+from typing import Optional
+
+from ..abstract import HashableModel
+from ..base import Chain
 
 
 class Encoding(str, Enum):
@@ -17,3 +21,21 @@ class MachineType(str, Enum):
 
     vm_instance = "vm-instance"
     vm_function = "vm-function"
+
+
+class PaymentType(str, Enum):
+    """Payment type for a program execution."""
+
+    hold = "hold"
+    stream = "stream"
+
+
+class Payment(HashableModel):
+    """Payment information for a program execution."""
+
+    chain: Chain
+    """Which chain to check for funds"""
+    receiver_address: Optional[str]
+    """Optional alternative address to send tokens to"""
+    payment_type: PaymentType
+    """Whether to pay by holding $ALEPH or by streaming tokens"""
